Split out slave-specific state from ConnectionContext into separate SlaveConnectionContext.
This change splits connection context objects for masters and slaves into two
separate classes: SlaveConnectionContext and MasterConnectionContext, rather 
than trying to have a single type play both roles. I think this is a bit cleaner,
given all the DCHECK(!is_slave()) and similar constraints in the old
ConnectionContext; but the primary motivation is to isolate the state specific
to slave connections, so I can refactor it not require SPDY-specific state.

r=mdsteele,bmcquade